requirements.txt
elevenlabs>=2.4.0
python-dotenv>=1.0.0
requests
flask
flask-cors 